Module 4 activity
Contains: tutorials.js, geojson.js, and my module 4 data (simply called
module4data.csv and module4.geojson). There are unfortunately also a
bunch of other files that aren't suppose to be here, but for some reason
where carried over when I created the repository. Last time I had issues
syncing up when I messed with the files that it wanted to carry over so
I'm leaving them for now with the hopes of cleaning house later on. The
tutorials.js seems to work fine as does geojson.js if I understood what
I was suppose to hand in correctly.