Module 5B assignment turn in
Contains main.js, style.css and index.html. The four main operators
work, though there is an issue with the size of my data that I have yet
to resolve. Due to the seemingly small increase my data goes through
over the years I discovered its currently difficult to see how the
symbols change as I use the slider or skip buttons. It does work, but
its hard to see, the most noticible difference seems to be the final
change on the slider. I ultimately could not get my filter to work
properly so I've temporary scrapped the code for it to insure that the
rest of the code worked properly. I felt it was better to hand in what I
have instead of handing it in late again.