Final Leaflet Lab
Commiting what I have done. Contains Index.html, main.js and style.css.
Unfortunately it does not contain the finished product I was hoping to
have at the end. There is no real style on the font, I was unable to
properly space the map title and I was unable to implement buttons that
used images in time for the due date.