Final Leaflet Lab
Due to issues syncing I had to rename style.css and main.js to
newstyle.css and newmain.css. Not sure why, but it wouldn't sync those
files. This commit includes newstyle.css, index.html, and newmain.js.
Unfortunately it does not contain the finished product I was hoping to
have at the end. There is no real style on the font, I was unable to
properly space the map title and I was unable to implement buttons that
used images in time for the due date.